+++ modules/PayButton/Dialog/PaymentComplete.js
