+++ modules/PayButton/Dialog/PaymentProgress.js
