+++ modules/PayButton/Dialog/index.js
