+++ modules/PayButton/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "PayButton",
+  "version": "0.1.0",
+  "main": "index.js",
+  "license": "MIT"
+}
