+++ modules/api/register.js
