+++ modules/website/.eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = require('../../.eslintrc.js')
