+++ modules/website/src/index.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import './index.css'
-import Site from './Site'
+import 'style/base.css'
+import 'style/main-content.css'
+import Site from 'Site'
 
 ReactDOM.render(<Site />, document.getElementById('root'))
