+++ modules/website/src/pages/IdentificationPage.js
@@ -1,18 +1,13 @@
-import React, {Component} from 'react'
-
+import React, { Component } from 'react'
 import TextField from '@material-ui/core/TextField'
 import Button from '@material-ui/core/Button'
-
-import './../MainContent.css'
-
-const bchaddr = require('bchaddrjs')
+import bchaddr from 'bchaddrjs'
 
 class IdentificationPage extends Component {
-
   state = {
     value: '',
     username: false,
-    loginError: false
+    loginError: false,
   }
 
   handleChange = () => {
@@ -21,14 +16,14 @@ class IdentificationPage extends Component {
       // check for a valid Bitcoin Cash address and convert if needed
       var valid = false
       try {
-        var isCash   = bchaddr.isCashAddress(value.trim())
+        var isCash = bchaddr.isCashAddress(value.trim())
         var isLegacy = bchaddr.isLegacyAddress(value.trim())
         if (isLegacy) {
           value = bchaddr.toCashAddress(value)
           isCash = true
         }
         valid = isCash
-        if (value.indexOf(':') === -1){
+        if (value.indexOf(':') === -1) {
           value = 'bitcoincash:' + value
         }
       } catch (e) {}
@@ -45,15 +40,21 @@ class IdentificationPage extends Component {
 
   toggleUsername = () => {
     if (this.state.username === false) {
-      this.setState({username: true})
+      this.setState({ username: true })
     } else {
-      this.setState({username: false})
+      this.setState({ username: false })
     }
   }
 
-  submitData (type, value) {
+  submitData(type, value) {
     var xhr = new XMLHttpRequest()
-    xhr.open('GET', 'https://gateway.cash/api/login?type=' + type + '&value=' + encodeURIComponent(value))
+    xhr.open(
+      'GET',
+      'https://gateway.cash/api/login?type=' +
+        type +
+        '&value=' +
+        encodeURIComponent(value),
+    )
     xhr.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded')
     xhr.onload = () => {
       if (xhr.status === 200) {
@@ -64,7 +65,9 @@ class IdentificationPage extends Component {
         } else if (response === 'register') {
           this.props.updateView('register')
         } else {
-          setTimeout(() => {this.setState({loginError: true})}, 5000)
+          setTimeout(() => {
+            this.setState({ loginError: true })
+          }, 5000)
         }
       } else {
         console.error('Username check request failed', xhr)
@@ -78,24 +81,38 @@ class IdentificationPage extends Component {
     this.handleChange()
   }
 
-  render () {
+  render() {
     return (
       <div className="mainContent">
-        <h1 className="mainHeading">Enter your {this.state.username ? 'gateway.cash username' : 'Bitcoin Cash address'}</h1>
-        <form onSubmit={this.handleSubmit} noValidate autoComplete="off" >
+        <h1 className="mainHeading">
+          Enter your{' '}
+          {this.state.username
+            ? 'gateway.cash username'
+            : 'Bitcoin Cash address'}
+        </h1>
+        <form onSubmit={this.handleSubmit} noValidate autoComplete="off">
           <TextField
             onChange={this.handleChange}
             style={{
-              width:'100%'
+              width: '100%',
             }}
             id="loginValue"
             placeholder={this.state.username ? 'username' : 'bitcoincash:q...'}
             autoFocus={true}
-            helperText={this.state.loginError ? 'Check your username or use your address instead' : ''}
+            helperText={
+              this.state.loginError
+                ? 'Check your username or use your address instead'
+                : ''
+            }
           />
-          <br/>
-          <br/>
-          <Button onClick={this.toggleUsername}>I have a {this.state.username ? 'Bitcoin Cash address' : ' gateway.cash username'}</Button>
+          <br />
+          <br />
+          <Button onClick={this.toggleUsername}>
+            I have a{' '}
+            {this.state.username
+              ? 'Bitcoin Cash address'
+              : ' gateway.cash username'}
+          </Button>
         </form>
       </div>
     )
