+++ modules/website/src/pages/LoginPage.js
@@ -1,21 +1,22 @@
-import React, {Component} from 'react'
-
+import React, { Component } from 'react'
 import TextField from '@material-ui/core/TextField'
 import Button from '@material-ui/core/Button'
 
-import './../MainContent.css'
-
 class LoginPage extends Component {
-
   state = {
-    loginError: false
+    loginError: false,
   }
 
   handleChange = () => {
     let value = document.getElementById('passwordValue').value
-    if (value.length < 12) { return }
+    if (value.length < 12) {
+      return
+    }
     var xhr = new XMLHttpRequest()
-    xhr.open('GET', 'https://gateway.cash/api/password?password='+encodeURIComponent(value))
+    xhr.open(
+      'GET',
+      'https://gateway.cash/api/password?password=' + encodeURIComponent(value),
+    )
     xhr.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded')
     xhr.onload = () => {
       if (xhr.status === 200) {
@@ -24,7 +25,9 @@ class LoginPage extends Component {
         if (response === 'ok') {
           this.props.updateView('createbutton')
         } else {
-          setTimeout(() => {this.setState({loginError: true})}, 5000)
+          setTimeout(() => {
+            this.setState({ loginError: true })
+          }, 5000)
         }
       } else {
         console.error('Password check request failed')
@@ -42,25 +45,27 @@ class LoginPage extends Component {
     this.props.updateView('forgot')
   }
 
-  render () {
+  render() {
     return (
       <div className="mainContent">
         <h1 className="mainHeading">Enter your password</h1>
-        <form onSubmit={this.handleSubmit} noValidate autoComplete="off" >
+        <form onSubmit={this.handleSubmit} noValidate autoComplete="off">
           <TextField
             style={{
-              width:'100%'
+              width: '100%',
             }}
             onChange={this.handleChange}
             id="passwordValue"
             type="password"
             placeholder="password..."
             autoFocus={true}
-            helperText={this.state.loginError ? 'Check your password and try again.' : ''}
+            helperText={
+              this.state.loginError ? 'Check your password and try again.' : ''
+            }
           />
         </form>
-        <br/>
-        <br/>
+        <br />
+        <br />
         <Button onClick={this.handleClick}>Forgot password</Button>
       </div>
     )
