--- src/paySite/pages/MainContent.css
@@ -1,69 +1,0 @@
-.mainContent {
-  align: center;
-  text-align: center;
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  -webkit-transform: translate(-50%, -50%);
-  -moz-transform: translate(-50%, -50%);
-  -o-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-}
-
-.mainHeading {
-  font-family: helvetica;
-  font-weight: bold;
-  font-size: 2em;
-}
-
-.logoImage {
-	width: 10em;
-}
-
-.container {
-  width: 90%;
-  margin-left: auto;
-  margin-right: auto;
-  font-family: helvetica;
-  margin-bottom: 1em;
-  align: left;
-  text-align: left;
-  word-wrap: break-word;
-}
-
-.sourceCode {
-  background-color: #dddddd;
-  color: black;
-  border: none;
-  outline: none;
-  font-family: monospace;
-  align: left;
-  text-align: left;
-  padding: 0.5em;
-  word-wrap: break-word;
-  box-shadow: 0.2em 0.2em 0.4em #99c;
-}
-
-.payment {
-	background-color: #dddddd;
-  color: black;
-  border: none;
-  outline: none;
-  align: left;
-  text-align: left;
-  padding: 0.5em;
-  word-wrap: break-word;
-  box-shadow: 0.2em 0.2em 0.4em #99c;
-  margin-bottom: 1em;
-  margin-top: 1em;
-}
-
-.payment a {
-	color: blue;
-	text-decoration: none;
-}
-
-.payment a:visited {
-	color: blue;
-	text-decoration: none;
-}
